add some lisa functions; put arguments into util.h
