+++ src/util.h
@@ -1,0 +1,106 @@
+#include "DFG.h"
+#include "CGRA.h"
+#include "Module.h"
+#include "Port.h"
+
+
+#include <iostream>
+#include <chrono>
+#include <math.h>
+#include <set>
+#include <ctype.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <unistd.h>
+
+#ifndef MAPPER_UTIL_H_
+#define MAPPER_UTIL_H_
+
+struct arguments
+{
+	string dfg_filename;
+	int xdim = -1;
+	int ydim = -1;
+	string PEType;
+	string json_file_name;
+	int userII = 0;
+	bool noMutexPaths=false;
+	int backtracklimit = 2; // for PathFinderMapper, do not set this for a high number.  
+	bool use_json = false;
+	int ndps = 1;
+	int maxiter = 30;  // for PathFinderMapper,
+	int max_hops = 4;  // for HyCUBE
+	int mapping_method = 0; // 0: PathFinderMapper, 1: SAMapper (SimulatedAnnealing),  HeuristicMapper will not be used
+};
+
+arguments parse_arguments(int argn, char *argc[])
+{
+	arguments ret;
+
+	int aflag = 0;
+	int bflag = 0;
+	char *cvalue = NULL;
+	int index;
+	int c;
+
+	opterr = 0;
+
+	while ((c = getopt(argn, argc, "d:x:y:t:j:i:eb:m:r:h:")) != -1)
+		switch (c)
+		{
+		case 'd':
+			ret.dfg_filename = string(optarg);
+			break;
+		case 'x':
+			ret.xdim = atoi(optarg);
+			break;
+		case 'y':
+			ret.ydim = atoi(optarg);
+			break;
+		case 't':
+			ret.PEType = string(optarg);
+			break;
+		case 'n':
+			ret.ndps = atoi(optarg);
+			break;
+		case 'j':
+			ret.json_file_name = string(optarg);
+			ret.use_json = true;
+			break;
+		case 'i':
+			ret.userII = atoi(optarg);
+			break;
+		case 'e':
+			ret.noMutexPaths = true;
+			break;
+		case 'b':
+			ret.backtracklimit = atoi(optarg);
+			break;
+		case 'r':
+			ret.maxiter = atoi(optarg);
+		case 'm':
+			ret.mapping_method = atoi(optarg);
+			break;
+		case 'h':
+			ret.max_hops = atoi(optarg);
+			break;
+		case '?':
+			if (optopt == 'c')
+				fprintf(stderr, "Option -%c requires an argument.\n", optopt);
+			else if (isprint(optopt))
+				fprintf(stderr, "Unknown option `-%c'.\n", optopt);
+			else
+				fprintf(stderr,
+						"Unknown option character `\\x%x'.\n",
+						optopt);
+		default:
+			abort();
+		}
+		return ret;
+};
+
+
+
+
+
+#endif /* LISAMAPPER_H_ */
